feat(open-as): Change to /breweryrecipes open as <player>
